Refactorings
- Generators namespace with rename of CodeGenerator -> DefaultCodeGenerator
- Moved items into files and new namespaces
- updated tests

This is kind of a breaking change since nothing is in the default namespace
but I think it is nice having multiple code generators.
