+++ src/Codez/Codez.csproj
@@ -26,4 +26,19 @@
       <PackageReference Include="System.Threading.Tasks.Extensions" Version="4.6.0-preview.18571.3" />
     </ItemGroup>
 
+
+    <ItemGroup>
+      <Compile Remove="artifacts\**" />
+    </ItemGroup>
+
+
+    <ItemGroup>
+      <EmbeddedResource Remove="artifacts\**" />
+    </ItemGroup>
+
+
+    <ItemGroup>
+      <None Remove="artifacts\**" />
+    </ItemGroup>
+
 </Project>
