+++ src/Codez/Listeners/BeforeAttemptEvent.cs
@@ -1,0 +1,12 @@
+namespace Codez.Listeners
+{
+    public class BeforeAttemptEvent
+    {
+        public BeforeAttemptEvent(int attempt)
+        {
+            Attempt = attempt;
+        }
+        
+        public int Attempt { get; }
+    }
+}
