+++ src/Codez/Listeners/IListener.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+
+namespace Codez.Listeners
+{
+    public interface IListener
+    {
+        Task OnBeforeAttempt(BeforeAttemptEvent @event);
+        Task OnAfterAttempt(AfterAttemptEvent @event);
+    }
+}
