+++ test/Tests/CodeGeneratorTests.cs
@@ -1,11 +1,11 @@
 ﻿
 using Codez;
-
+using Codez.Generators;
 using Xunit.Abstractions;
 
 namespace Tests
 {
-    public class CodeGeneratorTests : CodeGeneratorTestsBase<CodeGenerator>
+    public class CodeGeneratorTests : CodeGeneratorTestsBase<DefaultCodeGenerator>
     {
         public CodeGeneratorTests(ITestOutputHelper output) 
             : base(output)
