+++ test/Tests/CodeGeneratorTestsBase.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 
 using Codez;
 using Codez.Alphabets;
+using Codez.Generators;
 using Codez.StopWords;
 using Codez.Uniques;
