+++ test/Tests/TransformerTests.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Codez;
 using Codez.Alphabets;
+using Codez.Generators;
 using Codez.Transformers;
 using Xunit;
 using Xunit.Abstractions;
@@ -20,7 +21,7 @@ namespace Tests
         [Fact]
         public async Task Can_generate_and_transform_to_container_names()
         {
-            var generator = new CodeGenerator(
+            var generator = new DefaultCodeGenerator(
                 alphabet: new StringAlphabet("0123456789"),
                 transformer: new ContainerNamesTransformer()
             );
@@ -36,7 +37,7 @@ namespace Tests
         [Fact]
         public async Task Can_fail_to_transform()
         {
-            var generator = new CodeGenerator(
+            var generator = new DefaultCodeGenerator(
                 alphabet: new StringAlphabet("0123456789"),
                 transformer: new ContainerNamesTransformer()
             );
